api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.7
PyPDF2==3.0.1
faiss-cpu==1.7.4
numpy==1.24.3
python-multipart==0.0.6
pydantic==2.5.0
typing-extensions==4.8.0